chore(chat messages): move scripts to their own file
